[Feature] Add a new search bar
