+++ ui/app/src/App.tsx
@@ -14,7 +14,7 @@
 import { Box } from '@mui/material';
 
 import { useLocation } from 'react-router-dom';
-import Header from './components/Header';
+import Header from './components/Header/Header';
 import Footer from './components/Footer';
 import Router from './Router';
 import { SignInRoute, SignUpRoute } from './model/route';
